+++ src/main/java/com/example/dpm/auth/utils/SecurityUtil.java
@@ -1,22 +1,16 @@
 package com.example.dpm.auth.utils;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 
 import com.example.dpm.auth.exception.CustomException;
 import com.example.dpm.auth.exception.ErrorCode;
 import com.example.dpm.auth.models.UserPrincipal;
 
-import java.util.Optional;
-
-
 public class SecurityUtil {
     private SecurityUtil() {}
 
-    public static int getCurrentUserId() {
+    public static Long getCurrentUserId() {
 
         final Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
@@ -24,13 +18,13 @@ public class SecurityUtil {
             throw new CustomException(ErrorCode.UNAUTHORIZED);
         }
 
-        int userId;
+        Long userId;
         if (authentication.getPrincipal() instanceof UserPrincipal userPrincipal) {
             userId = userPrincipal.getId();
         } else {
             throw new CustomException(ErrorCode.BAD_REQUEST);
         }
 
-        return userId;
+        return (Long) userId;
     }
 }
